pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-pixelator"
version = "1.0.0"
description = "Convert images to pixelated blocks with binary (1/0) patterns"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0"
]
keywords = ["image", "pixelator", "filter", "pattern", "binary", "dithering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

[project.urls]
Homepage = "https://github.com/yourusername/image-pixelator"
Repository = "https://github.com/yourusername/image-pixelator.git"
Issues = "https://github.com/yourusername/image-pixelator/issues"
Documentation = "https://github.com/yourusername/image-pixelator#readme"
